requirements.txt
fastapi
uvicorn
typing
sqlalchemy
sqlalchemy_utils
pydantic
python-dotenv
psycopg2